refactor(08-export): pdfCanvasRenderer を責務別モジュールへ分割
1026行の単一ファイル pdfCanvasRenderer.ts を、責務別の7モジュール
（pdfCanvasRenderer/ ディレクトリ）へ分割。ロジックは変更せずコードを
そのまま移設。消費者の import パス（utils/pdfCanvasRenderer）は index.ts
バレルに解決されるため無変更。

- types.ts: 共有4型（ScoringDataForPdf 他）
- annotationRenderer.ts: DrawingElement(内部型)・変換・drawElement
- markTinting.ts: getTintedMark + キャッシュ
- markPosition.ts: マーク/点数の位置計算
- scoreDrawing.ts: マーク画像・点数テキスト描画
- renderAnswerSheet.ts: 公開2関数（描画orchestration・プリロード）
- index.ts: バレル（公開API surface）

型規約に準拠: 複数ファイルで使用する型は types.ts へ集約、内部専用型
DrawingElement はローカル宣言、import type を使用、any/as の追加なし。
